Add organization model
Organization is a group of groups. Currently this is a very simple
implementation with attributes logo and name. The name is the name
of the orgnaization and the logo is expected to be an svg. A cap
was put on the size of the logo in chars simply to prevent a
malliciously large upload. Also verify logo is an svg and is html
parseable.